update metric interface

+++ fvm/environment/mock/environment.go
@@ -477,7 +477,7 @@ func (_m *Environment) DeductTransactionFees(payer flow.Address, inclusionEffort
 }
 
 // EVMBlockExecuted provides a mock function with given fields: txCount, totalGasUsed, totalSupplyInFlow
-func (_m *Environment) EVMBlockExecuted(txCount int, totalGasUsed uint64, totalSupplyInFlow uint64) {
+func (_m *Environment) EVMBlockExecuted(txCount int, totalGasUsed uint64, totalSupplyInFlow float64) {
 	_m.Called(txCount, totalGasUsed, totalSupplyInFlow)
 }
