+++ fvm/environment/programs_test.go
@@ -897,7 +897,7 @@ func (m *metricsReporter) SetNumberOfDeployedCOAs(count uint64) {}
 
 func (m *metricsReporter) EVMTransactionExecuted(_ uint64, _ bool, _ bool) {}
 
-func (m *metricsReporter) EVMBlockExecuted(_ int, _ uint64, _ uint64) {}
+func (m *metricsReporter) EVMBlockExecuted(_ int, _ uint64, _ float64) {}
 
 func (m *metricsReporter) RuntimeTransactionProgramsCacheMiss() {
 	m.CacheMisses++
