+++ module/metrics/noop.go
@@ -201,7 +201,7 @@ func (nc *NoopCollector) RuntimeTransactionProgramsCacheMiss()                  
 func (nc *NoopCollector) RuntimeTransactionProgramsCacheHit()                                   {}
 func (nc *NoopCollector) SetNumberOfDeployedCOAs(_ uint64)                                      {}
 func (nc *NoopCollector) EVMTransactionExecuted(_ uint64, _ bool, _ bool)                       {}
-func (nc *NoopCollector) EVMBlockExecuted(_ int, _ uint64, _ uint64)                            {}
+func (nc *NoopCollector) EVMBlockExecuted(_ int, _ uint64, _ float64)                           {}
 func (nc *NoopCollector) ScriptExecuted(dur time.Duration, size int)                            {}
 func (nc *NoopCollector) ScriptExecutionErrorLocal()                                            {}
 func (nc *NoopCollector) ScriptExecutionErrorOnExecutionNode()                                  {}
